Refactor some movement code
